2025-04-17-周四，CRC-7061
1. 修改了使用9脚时的参考电压，改为 4.2V内部参考电压
2. 修改了一个使用浮点类型的变量，改为u16格式
3. 修改了部分ad检测的函数，加入了滤波

+++ User/pwm.h
@@ -18,7 +18,8 @@ enum
 };
 
 void pwm_init(void);
-void _My_Adjust_Pwm(float Val);
+// void _My_Adjust_Pwm(float Val);
+void _My_Adjust_Pwm(u16 Val);
 void set_pwm_duty(void);
 void Adaptive_Duty(void);
